[MAJ] Classes CSS delete, edit et syndication renomm?es en pbt-icon-delete, pbt-icon-edit et pbt-icon-syndication Pr?c?demment la classe css delete afficher automatiquement une confirmation de suppression, ce n'est plus le cas, maintenant il faut utiliser data-confirmation="delete-element" Si vous souhaitez int?grer un message de confirmation simplement de vos codes, vous pouvez utiliser data-confirmation="votre message" Les boutons pour faire du tri dans les tableaux sont renomm? en pbt-icon-table-sort-up et pbt-icon-table-sort-down Ajout des classes CSS pbt-icon-plus et pbt-icon-minus (d'autres viendront)
Je pense que maintenant nous avons quelque chose de beaucoup plus coh?rent ;)

+++ articles/templates/AdminManageArticlesController.tpl
@@ -61,10 +61,10 @@
 				<span class="smaller">{articles.PUBLISHED_DATE}</span>
 			</td>
 			<td> 
-				<a href="{articles.U_EDIT_ARTICLE}" title="{articles.L_EDIT_ARTICLE}" class="edit"></a>
+				<a href="{articles.U_EDIT_ARTICLE}" title="{articles.L_EDIT_ARTICLE}" class="pbt-icon-edit"></a>
 			</td>
 			<td>
-				<a href="{articles.U_DELETE_ARTICLE}" title="{articles.L_DELETE_ARTICLE}" class="delete"></a>
+				<a href="{articles.U_DELETE_ARTICLE}" title="{articles.L_DELETE_ARTICLE}" class="pbt-icon-delete" data-confirmation="delete-element"></a>
 			</td>
 		</tr>
 		# END articles #
